requirements.txt
gradio_client
python-dotenv
Pillow
requests